+++ app/src/main/java/com/github/se/cyrcle/ui/authentication/SignInScreen.kt
@@ -33,6 +33,7 @@ import com.github.se.cyrcle.ui.navigation.Screen
 import com.github.se.cyrcle.ui.navigation.TopLevelDestinations
 import com.github.se.cyrcle.ui.theme.ColorLevel
 import com.github.se.cyrcle.ui.theme.atoms.Button
+import com.github.se.cyrcle.ui.theme.atoms.GoogleSignInButton
 import com.github.se.cyrcle.ui.theme.atoms.Text
 
 @Composable
